pyproject.toml
[project]
name = "dharideepa"
version = "0.1.0"
description = ""
authors = [
    {name = "siddhanth"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.26.0"

